Changed X, Y, stepX, and stepY values from ints to doubles for improved ball / bat collision physics Changed ball display from square to circle Improved ball / bat collision physics - Game now almost playable :)
TODO improve ball / bat collision physics further

+++ Breakout/src/com/gmail/jameshealey1994/breakout/GamePanel.java
@@ -47,7 +47,7 @@ public class GamePanel extends JPanel implements Runnable {
                 synchronized (Lock.LOCK) {
                     for (int i = 0; i < 10; i++) {
                         bat.clear();
-                        final int newPosX = bat.getX() - 1;
+                        final double newPosX = bat.getX() - 1;
                         if (bat.isValidMove(newPosX)) {
                             bat.setX(newPosX);
                         }
@@ -66,7 +66,7 @@ public class GamePanel extends JPanel implements Runnable {
                 synchronized (Lock.LOCK) {
                     for (int i = 0; i < 10; i++) {
                         bat.clear();
-                        final int newPosX = bat.getX() + 1;
+                        final double newPosX = bat.getX() + 1;
                         if (bat.isValidMove(newPosX)) {
                             bat.setX(newPosX);
                         }
